refactor: LottoService 삭제
## why
불필요한 서비스 계층을 두지 않고 도메인 내에서 문제를 해결하고자 LottoService는 삭제하였으며 Controller에서 도메인에 구현된 메서드를 callback하여 로직을 해결하고자 하였습니다.
이에 따라 전체 구조를
[starter] -> [viewer] <-> [contoller]
로 변경하였습니다.
