+++ src/main/java/kr/codesquad/LottoController.java
@@ -1,26 +1,36 @@
 package kr.codesquad;
 
 import kr.codesquad.domain.Lotto;
+import kr.codesquad.domain.Row;
 import kr.codesquad.domain.Statistic;
 import kr.codesquad.domain.WinningNumbers;
 
+import java.util.List;
+
 public class LottoController {
-    private final LottoService lottoService;
+    private final AutoLottoGenerator autoLottoGenerator;
 
     public LottoController() {
-        this.lottoService = new LottoService();
+        this.autoLottoGenerator = new AutoLottoGenerator();
     }
 
-    public Lotto receiveInput(int inputMoney) {
+    public Lotto createLotto(int inputMoney) {
         if (inputMoney < Statistic.SINGLE_PRICE) {
             throw new IllegalArgumentException("금액 1000미만임.");
         }
-        //랜덤 숫자, 수동 숫자 결정
-        return lottoService.receiveRandomRows(inputMoney);
+        //todo: 랜덤 숫자, 수동 숫자 결정
+        List<Row> rows = autoLottoGenerator.generateRows(inputMoney);
+        return Lotto.createLotto(rows, inputMoney);
     }
 
     public Statistic result(Lotto lotto, WinningNumbers winningNumbers) {
         lotto.compareLotto(winningNumbers);
-        return lottoService.getResult(lotto);
+        return getResult(lotto);
+    }
+
+    public Statistic getResult(Lotto lotto) {
+        Statistic statistic = new Statistic(lotto.getInputMoney());
+        statistic.calculate(lotto.getTotalLotto());
+        return statistic;
     }
 }
